# backend/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "description": "The main API structure\n\n- `/api/v1beta/` <- Most of the API is nested under here. All of the resources there are scoped to what is accessible by the authenticated identity.\n- `/api/v1beta/me` <- Everything under this path is scoped to the subject of the authenticated identity.\nThis means that the identity may be authorized to view more resources, but this is the default view for them.\nE.g. the chats route scoped under there will only list the chats created by the user, but the user may be authorized to also view chats shared by other users.\n",
    "version": ""
  },
  "paths": {
    "/api/v1beta/assistants": {
      "get": {
        "tags": [
          "assistants"
        ],
        "summary": "List all assistants available to the user",
        "operationId": "list_assistants",
        "responses": {
          "200": {
            "description": "Successfully retrieved user's assistants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Assistant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "assistants"
        ],
        "summary": "Create a new assistant",
        "operationId": "create_assistant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successfully created the assistant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantWithFiles"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1beta/assistants/{assistant_id}": {
      "get": {
        "tags": [
          "assistants"
        ],
        "summary": "Get a specific assistant with its files",
        "operationId": "get_assistant",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "description": "The ID of the assistant to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the assistant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantWithFiles"
                }
              }
            }
          },
          "400": {
            "description": "Invalid assistant ID format"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "404": {
            "description": "Assistant not found or access denied"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "assistants"
        ],
        "summary": "Update an existing assistant",
        "operationId": "update_assistant",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "description": "The ID of the assistant to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated the assistant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantWithFiles"
                }
              }
            }
          },
          "400": {
            "description": "Invalid assistant ID format or request data"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "404": {
            "description": "Assistant not found or access denied"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1beta/assistants/{assistant_id}/archive": {
      "post": {
        "tags": [
          "assistants"
        ],
        "summary": "Archive an assistant",
        "operationId": "archive_assistant",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "description": "The ID of the assistant to archive",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveAssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully archived the assistant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveAssistantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid assistant ID format"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "404": {
            "description": "Assistant not found or access denied"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1beta/files/{file_id}": {
      "get": {
        "tags": [],
        "summary": "Get a single file by its ID",
        "description": "This endpoint retrieves information about a specific file by its ID.",
        "operationId": "get_file",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "description": "The ID of the file to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadItem"
                }
              }
            }
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "404": {
            "description": "When the file doesn't exist or doesn't belong to the user"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1beta/share-grants": {
      "get": {
        "tags": [
          "share_grants"
        ],
        "summary": "List share grants for a resource",
        "operationId": "list_share_grants",
        "parameters": [
          {
            "name": "resource_type",
            "in": "query",
            "description": "The type of resource (e.g., 'assistant')",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "query",
            "description": "The ID of the resource",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved share grants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListShareGrantsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "403": {
            "description": "User does not own the resource"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "share_grants"
        ],
        "summary": "Create a new share grant",
        "operationId": "create_share_grant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successfully created the share grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateShareGrantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "403": {
            "description": "User does not own the resource"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveAssistantRequest": {
        "type": "object",
        "description": "Request to archive an assistant"
      },
      "ArchiveAssistantResponse": {
        "type": "object",
        "description": "Response when archiving an assistant",
        "required": [
          "id",
          "archived_at"
        ],
        "properties": {
          "archived_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the assistant was archived"
          },
          "id": {
            "type": "string",
            "description": "The ID of the archived assistant"
          }
        }
      },
...
      "Assistant": {
        "type": "object",
        "description": "An assistant model",
        "required": [
          "id",
          "name",
          "prompt",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "archived_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this assistant was archived"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this assistant was created"
          },
          "default_chat_provider": {
            "type": "string",
            "description": "Default chat provider/model ID for this assistant"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the assistant"
          },
          "id": {
            "type": "string",
            "description": "The unique ID of the assistant"
          },
          "mcp_server_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of MCP server IDs available to this assistant"
          },
          "name": {
            "type": "string",
            "description": "The display name of the assistant"
          },
          "prompt": {
            "type": "string",
            "description": "The system prompt used by the assistant"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this assistant was last updated"
          }
        }
      },
...
      "AssistantWithFiles": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Assistant",
            "description": "The assistant information"
          },
          {
            "type": "object",
            "required": [
              "files"
            ],
            "properties": {
              "files": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AssistantFile"
                },
                "description": "Files associated with this assistant"
              }
            }
          }
        ],
        "description": "An assistant with its associated files"
      },
...
    }
  }
}